Ability to add new sources via modal implemented.
Adding new sources required a new user interface component to go along with it, thus a modal. A modal should be able to be dynamically loaded and then disposed of after use, as it is not a core component of the program. Thus a bit of custom templating was also implemented for this case.